\name{PETsample}
\alias{PETsample}
\title{Data from DICOM Image File}
\description{
  The 'PETsample' dataset contains the data extracted from the DICOM image "003_S_1059.img".
}
\usage{
  data("PETsample")
}
\format{
  A PET neuroimaging file with dimensions and other metadata stored in the ".hdr" file.
}
\source{
  This file has been obtained with permission from ADNI.
}
\keyword{datasets}
