% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseCreator.R
\name{databaseCreator}
\alias{databaseCreator}
\title{Create a database of processed PET image data with the appropiate format}
\usage{
databaseCreator(pattern)
}
\arguments{
\item{pattern}{\code{character}, a regular expression pattern that specifies which files to process.}
}
\value{
A \code{data.frame} that aggregates processed data from each image. Each row represents data from one image, including subject numbers and image data.
}
\description{
This function automates the processing of PET images based on a specified file name pattern within a working directory.
It reads each file matching the pattern, processes it using \code{neuroSCC::neuroCleaner}, and compiles the results into a comprehensive data frame.
}
\details{
The function first checks if there are files matching the pattern in the working directory.
If no files are found, it throws an error. Otherwise, it processes each file individually, extracts necessary data,
and appends it to a growing database. The function leverages \code{neuroSCC::neuroCleaner} for data processing.
Each file's subject number is extracted from its name using regular expressions.
}
\examples{
# Set the working directory where your PET images are stored
setwd("~/GitHub/PhD-2023-Neuroimage-article-SCC-vs-SPM/PETimg_masked for simulations")

# Define the pattern for file names to process
pattern <- "^masked_swwwC\\\\d+_tripleNormEsp_w00_rrec_OSEM3D_32_it1.nii"

# Create the database
database_CN <- databaseCreator(pattern)

}
\seealso{
\link[neuroSCC]{neuroCleaner} for the underlying image processing.
}
