% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDimensions.R
\name{getDimensions}
\alias{getDimensions}
\title{Get Dimensions from a Neuroimaging File}
\usage{
getDimensions(file)
}
\arguments{
\item{file}{A NIFTI file object or filename to extract dimensions from.}
}
\value{
A list containing:
\itemize{
  \item \code{xDim}: Number of voxels in the X dimension
  \item \code{yDim}: Number of voxels in the Y dimension
  \item \code{zDim}: Number of slices in the Z dimension
  \item \code{dim}: Total number of voxels (xDim * yDim)
}
}
\description{
Extracts dimensional information from NIFTI or similar neuroimaging files.
This function is designed to work together with \code{neuroCleaner()}
but can also be used independently for informative purposes.
}
\details{
The function can handle both NIFTI file paths and pre-loaded oro.nifti file objects.
It provides a consistent way to extract dimensional information across the package.
}
\examples{
\dontrun{
# Using a file path
dims <- neuroSCC::getDimensions("path/to/your/image.nii")

# If used within neuroCleaner
file <- oro.nifti::readNIfTI("path/to/your/image.nii")
dims <- neuroSCC::getDimensions(file)
}

}
