% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDimensions.R
\name{getDimensions}
\alias{getDimensions}
\title{Get Dimensions from a DICOM File}
\usage{
getDimensions(filename = NULL)
}
\arguments{
\item{filename}{\code{character}, optional; the name of the DICOM file to read.
If not provided, the function will search for the first \code{.img} file in the current working directory.}
}
\value{
A list containing \code{xDim}, \code{yDim}, and \code{dim}, representing the dimensions of the image in the X and Y axes,
and the total number of elements in the image, respectively.
}
\description{
This function loads a DICOM image file using the \code{oro.nifti} package and extracts its dimensions.
It provides the X and Y dimensions as well as the total number of elements ('dim') in the image data.
}
\details{
If no filename is provided, the function searches the current directory for the first file with a \code{.img} extension.
It stops with an error message if no such files are found. It is important to ensure that the specified file or files in the directory are in the DICOM format.
}
\examples{
# If 'filename' is not provided, it will get the first image in the current working directory:
dimensions <- getDimensions()

# Providing a specific filename:
dimensions <- getDimensions("003_S_1059.img")
}
\seealso{
\link[oro.nifti]{readNIfTI} for the function used to read the DICOM files.
}
