% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixCreator.R
\name{matrixCreator}
\alias{matrixCreator}
\title{Convert PET image data to a functional data matrix format}
\usage{
matrixCreator(database, pattern, param.z, xy)
}
\arguments{
\item{database}{A data frame containing the PET image data with columns for 'CN_number', 'z', 'x', 'y', and 'pet' values, created by \code{databaseCreator}.}

\item{pattern}{The regular expression pattern to match filenames in the database. This pattern should correspond to the naming conventions of the PET image files you want to process.}

\item{param.z}{The specific z-coordinate slice to analyze. This parameter should be defined prior to running this function if not included in the script that calls it.}

\item{xy}{The total number of data points per z-slice, calculated as the product of the dimensions x and y of the image slice. If not predefined, it must be calculated by the user based on the image dimensions.}
}
\value{
A matrix where each row represents the PET data from one control subject, formatted as a continuous line of data points to simulate a function.
}
\description{
This function transforms PET image data, previously processed into a database format by \code{databaseCreator}, into a matrix format suitable for functional data analysis. Each row of the matrix represents a function, corresponding to data from one patient or control subject.
}
\details{
\code{matrixCreator} follows \code{databaseCreator} in the data processing workflow. It reads the data for each file specified by the matched pattern and extracts only the data for the specified z-slice. The function ensures that each row in the resulting matrix corresponds to one patient or control subject, transforming each subset of data into a format suitable for functional data analysis.
}
\examples{
# Assuming 'database_CN', 'pattern', 'param.z', and 'xy' are already defined
SCC_CN <- matrixCreator(database_CN, pattern, param.z, xy)

}
\seealso{
\link[databaseCreator]{databaseCreator} and \link[neuroCleaner]{neuroCleaner} for initial data processing and image cleaning.
}
