% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanNormalization.R
\name{meanNormalization}
\alias{meanNormalization}
\title{Mean Average Normalization for SCC Data}
\usage{
meanNormalization(SCC_data)
}
\arguments{
\item{SCC_data}{\code{matrix}, a matrix where each row represents SCC data to be normalized.}
}
\value{
A \code{matrix} where each row has been normalized by its mean value.
}
\description{
This function normalizes each row of the given SCC data by its mean value.
It divides each element in a row by the mean of that row, ignoring NA values.
}
\details{
The function iterates over each row of the SCC data matrix, calculates the mean of the row,
and then divides each element of the row by the calculated mean.
}
\examples{
# Assume SCC_data is a matrix of SCC values
normalized_SCC_data <- neuroSCC::meanNormalization(SCC_data)

}
