% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroContour.R
\name{neuroContour}
\alias{neuroContour}
\title{Obtain and save neuroimaging contours}
\usage{
neuroContour(data, levels = c(0))
}
\arguments{
\item{data}{\code{data.frame}, a data frame containing the neuroimaging data to process.}

\item{levels}{\code{numeric}, a vector of levels at which to draw the contours. Default is \code{c(0)}.}
}
\value{
A \code{list} of data frames, where each data frame contains the x and y coordinates of a contour.
}
\description{
This function extracts contours from neuroimaging data where values change according to specified levels.
It uses the \code{contoureR::getContourLines} function to obtain the contours and stores the coordinates in a list.
}
\details{
The function filters the contours by their \code{GID} and stores the coordinates of each contour in a list.
It ensures that the \code{contoureR} package is loaded before attempting to use its functions.
}
\examples{
# Example usage:
# Load sample data
data <- some_neuroimaging_data

# Get contours at level 0
contours <- neuroContour(data, levels = c(0))

# Plot the first contour
plot(contours[[1]])
if (length(contours) > 1) {
  for (j in 2:length(contours)) {
    points(contours[[j]]) # Holes or internal contours
  }
}

}
\seealso{
\link[contoureR]{getContourLines} for the underlying contour extraction.
}
