% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroSCC-package.R
\docType{package}
\name{neuroSCC-package}
\alias{neuroSCC}
\alias{neuroSCC-package}
\title{neuroSCC: Estimation of SCC's for PET Neuroimaging Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides functions for calculating Simultaneous Confidence Corridors (SCCs) on PET neuroimaging data. It is designed to help with the replication of methods implemented in the Ph.D. Thesis "Development of statistical methods for neuroimage data analysis towards early diagnostic of neurodegenerative diseases" and it is best followed using the scripts available in the link below. Further information on the methodology for SCCs estimation can be found in: "Wang Y, Wang G, Wang L, Ogden RT. Simultaneous confidence corridors for mean functions in functional data analysis of imaging data. *Biometrics*. 2020 Jun;76(2):427-437".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iguanamarina/neuroSCC}
}

}
\author{
\strong{Maintainer}: Juan A. Arias \email{iguanamarina@protonmail.com}

}
\keyword{internal}
