% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroSCC-package.R
\docType{package}
\name{neuroSCC-package}
\alias{neuroSCC}
\alias{neuroSCC-package}
\title{neuroSCC: Estimation of SCC's for PET Neuroimaging Data}
\description{
Este paquete proporciona funciones auxiliares para calcular Corredores de Confianza Simultáneos (SCCs)
en datos de neuroimagen PET. Incluye funciones para cargar datos neuroimagen y demográficos en el formato estándar requerido,
y asiste en varias tareas a lo largo del proceso. Este paquete está diseñado para ayudar con el proceso
de replicación de un artículo fundamental de mi tesis "Development of statistical methods for neuroimage data analysis
towards early diagnostic of neurodegenerative diseases" y se sigue mejor utilizando los scripts disponibles en
\url{https://github.com/iguanamarina/SCCneuroimage}. Este sirve como un paquete de soporte.
\figure{figures/logo.png}{options: width=120}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iguanamarina/neuroSCC}
}

}
\author{
\strong{Maintainer}: Juan A. Arias Developer \email{iguanamarina@protonmail.com}

}
\keyword{internal}
