% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroSCC-package.R
\docType{package}
\name{neuroSCC-package}
\alias{neuroSCC}
\alias{neuroSCC-package}
\title{neuroSCC: Bridging Simultaneous Confidence Corridors With Applications and Performance Evaluation in PET 'Neuroimaging'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions for the structured processing of PET 'Neuroimaging' data in preparation for the estimation of 'Simultaneous Confidence Corridors (SCCs)' for one-group, two-group, or single-patient vs group comparisons. The package facilitates PET image loading, data restructuring, integration into a Functional Data Analysis framework, contour extraction, post-SCC identification of significant results, and performance evaluation. It bridges established neuroimaging packages (e.g., 'oro.nifti') with novel statistical methodologies (i.e., 'ImageSCC') and enables reproducible analysis pipelines, including comparison with Statistical Parametric Mapping ('SPM').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iguanamarina.github.io/neuroSCC/}
  \item \url{https://github.com/iguanamarina/PhD-2023-SCC-vs-SPM-Group-vs-Group}
  \item \url{https://github.com/iguanamarina/PhD-2024-SCC-vs-SPM-SinglePatient-vs-Group}
  \item Report bugs at \url{https://github.com/iguanamarina/neuroSCC/issues}
}

}
\author{
\strong{Maintainer}: Juan A. Arias Lopez \email{juanantonio.arias.lopez@usc.es} (\href{https://orcid.org/0000-0002-3355-6393}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Virgilio Gomez Rubio \email{Virgilio.Gomez@uclm.es} (\href{https://orcid.org/0000-0002-4791-3072}{ORCID}) [reviewer]
  \item Pablo Aguiar Fernandez \email{pablo.aguiar@usc.es} (\href{https://orcid.org/0000-0002-7322-2195}{ORCID}) [thesis advisor]
  \item Andrew Haddon Kemp \email{A.H.Kemp@swansea.ac.uk} (\href{https://orcid.org/0000-0003-1146-3791}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
