% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processROIs.R
\name{processROIs}
\alias{processROIs}
\title{Process ROIs and Save Data Tables}
\usage{
processROIs(base_dir, regions, numbers)
}
\arguments{
\item{base_dir}{The base directory where the ROI files and tables are located.}

\item{regions}{A vector of region names to be processed.}

\item{numbers}{A vector of numbers corresponding to the patient or control subjects.}
}
\value{
This function does not return a value. It performs data processing and saves the resulting tables to the specified directory.
}
\description{
This function processes regions of interest (ROIs) from PET image data and saves the resulting data tables. It checks if the ROI tables already exist and, if not, processes the data using \code{neuroSCC::neuroCleaner} to clean and structure the data.
}
\details{
\code{processROIs} iterates over each combination of region and number, checking if the corresponding ROI table file already exists. If the file does not exist, the function processes the data using \code{neuroSCC::neuroCleaner}, adds necessary metadata, and saves the cleaned data as an RDS file. The function provides informative messages about the progress of processing and file saving.
}
\examples{
# Assuming 'base_dir', 'regions', and 'numbers' are already defined
processROIs(base_dir, regions, numbers)

}
